+++ __tests__/api/robots.test.js
@@ -23,7 +23,7 @@ describe("robots api", () => {
     });
     await handler(req, res);
     expect(res._getData()).toBe(
-      "User-agent: *\nDisallow: /api\nDisallow: /projects/*\nDisallow: /confirmation.js\nDisallow: /thankyou.js\n"
+      "User-agent: *\nDisallow: /api\nDisallow: /projects/*\nDisallow: /confirmation.js\nDisallow: /thankyou.js\nDisallow: /notsupported.js\n"
     );
   });
 });