+++ components/molecules/CopyToClipboard.test.js
@@ -1,0 +1,34 @@
+/**
+ * @jest-environment jsdom
+ */
+import { render, screen, fireEvent } from "@testing-library/react";
+import "@testing-library/jest-dom/extend-expect";
+import { Primary } from "./CopyToClipboard.stories";
+import { axe, toHaveNoViolations } from "jest-axe";
+
+expect.extend(toHaveNoViolations);
+
+describe("CopyToClipboard", () => {
+  let mockFn;
+  beforeEach(() => {
+    mockFn = jest.fn();
+  });
+  afterEach(() => {
+    mockFn.mockRestore();
+  });
+  it("renders component correctly", () => {
+    render(<Primary {...Primary.args} onChange={mockFn} />);
+    const inputElem = screen.getByTestId("clipboard-controlled");
+    fireEvent.change(inputElem, { target: { value: "h" } });
+    expect(inputElem.value).toBe("");
+    expect(mockFn.mock.calls.length).toBe(1);
+    expect(mockFn.mock.calls[0][0]).toBe("h");
+    screen.getByText(Primary.args.buttonText);
+  });
+
+  it("has no accessibility violations", async () => {
+    const { container } = render(<Primary {...Primary.args} />);
+    const results = await axe(container);
+    expect(results).toHaveNoViolations();
+  });
+});